requirements.txt
hvac 
psycopg2-binary 
python-consul
